# Use arrowutils.Take to filter records