Closes #672