@gernest 2024-01-22
I will remove `buildIndexRanges` after this being merged since we no longer use it. My eyeball test expects significan improvement in throughput and memory .

I will create a simple benchmark, but I was hoping someone who can run `bench_test.go` benchmarks can help here to get a   better picture. I can't run the benchmarks because they need `corpus` from `parca` that I don't have access to.