@thorfour 2024-01-22
```bash
name            old time/op    new time/op    delta
Query/Merge-10     128µs ± 4%     123µs ± 2%  -3.99%  (p=0.001 n=10+8)
Query/Range-10     128µs ± 3%     131µs ± 7%    ~     (p=0.400 n=9+10)

name            old alloc/op   new alloc/op   delta
Query/Merge-10     186kB ± 0%     186kB ± 0%    ~     (p=0.090 n=9+10)
Query/Range-10     186kB ± 0%     186kB ± 0%    ~     (p=0.898 n=10+10)

name            old allocs/op  new allocs/op  delta
Query/Merge-10     1.40k ± 0%     1.40k ± 0%    ~     (all equal)
Query/Range-10     1.41k ± 0%     1.41k ± 0%    ~     (all equal)
```

ran some of the benchmarks that I could get working